add PI mutex tests
for robust mutex tests, the code was already setup to be run more than
once with different attributes (pshared); I've added PI the same way.

for the main mutex tests, it would take a major refactoring to test
and report errors well for both the normal and PI cases, so I just
duplicated the source file with PI hard-coded on. this could be
improved later if desired.