add utime (utimensat, futimens) functional tests
these tests check for the ability to set file timestamps using
UTIME_NOW, UTIME_OMIT, and explicit timespecs. they indirectly check
that fstat works as well, and include a check of the first Y2038
timestamp that overflows 32-bit time_t, reporting errors for a Y2038
EOL implementaton (if time_t is 32-bit) or a kernel, filesystem, or
library implementation that fails to set and read back timestamps past
the 32-bit limit despite having a time_t type that can represent such
a timestamp.