fix flockfile-list regression test
this test should check if freed stdio memory is clobbered
after an funlockfile.

the old method was very fragile: it tried to reuse the
freed memory via a malloc and check if that allocation is
clobbered. now musl supports malloc interposition so we
can use that to directly check the required property.